// src/app/business-logic/model/tasks/configurationItem.ts
/**
 * tasks
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.13
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



export interface ConfigurationItem {
    /**
     * Name of the parameter. Should be unique
     */
    name?: string;
    /**
     * Value of the parameter
     */
    value?: string;
    /**
     * Type of the parameter. Optional
     */
    type?: string;
    /**
     * The parameter description. Optional
     */
    description?: string;
}